% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CohortMatching-package.R
\docType{package}
\name{CohortMatching-package}
\alias{CohortMatching}
\alias{CohortMatching-package}
\title{CohortMatching: Create Matched Generated Cohorts in the OMOP Common Data Model}
\description{
This package provide functions and tools to create matched cohorts using the OMOP Common Data Model.
}
\author{
\strong{Maintainer}: Marta Alcalde-Herraiz \email{marta.alcaldeherraiz@ndorms.ox.ac.uk}

Authors:
\itemize{
  \item Marti Catala \email{marti.catalasabate@ndorms.ox.ac.uk} (\href{https://orcid.org/0000-0003-3308-9905}{ORCID})
}

}
\keyword{internal}
