% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateMatchedCohort.R
\name{generateMatchedCohort}
\alias{generateMatchedCohort}
\title{generateMatchedCohort summary}
\usage{
generateMatchedCohort(
  cdm,
  name,
  targetCohortName,
  targetCohortId = NULL,
  matchSex = TRUE,
  matchYearOfBirth = TRUE,
  ratio = 1
)
}
\arguments{
\item{cdm}{cdm object}

\item{name}{name of the new cohort}

\item{targetCohortName}{Name of the target cohort to match}

\item{targetCohortId}{cohort_definition_id of interest. If it is NULL, all the cohort_definition_id present in the cohort will be matched}

\item{matchSex}{can be TRUE or FALSE}

\item{matchYearOfBirth}{can be TRUE or FALSE}

\item{ratio}{at the moment, it is not used this argument}
}
\value{
plot
}
\description{
generateMatchedCohort summary
}
